+++ src/main/java/me/srrapero720/watermedia/api/url/UrlAPI.java
@@ -16,7 +16,7 @@ import java.util.*;
 import static me.srrapero720.watermedia.WaterMedia.LOGGER;
 
 public class UrlAPI {
-    private static final Marker IT = MarkerManager.getMarker("URLApi");
+    public static final Marker IT = MarkerManager.getMarker("UrlAPI");
     private static final List<URLFixer> FIXERS = new ArrayList<>();
 
     /**